promote: convert check_assets tests to node:test
Refs: https://github.com/nodejs/build/issues/3958
